Cleanup. Floats to double.

+++ platform/unix/scaffold/graphics.d
@@ -544,7 +544,7 @@ void createBitmapBrush(BrushPlatformVars* brush, ref ViewPlatformVars viewVarsSr
 	Cairo.cairo_pattern_set_extend(brush.handle, Cairo.cairo_extend_t.CAIRO_EXTEND_REPEAT);
 }
 
-void createGradientBrush(BrushPlatformVars* brush, float origx, float origy, float[] points, Color[] clrs, float angle, float width) {
+void createGradientBrush(BrushPlatformVars* brush, double origx, double origy, double[] points, Color[] clrs, double angle, double width) {
 	double x0, y0, x1, y1;
 	x0 = origx;
 	y0 = origy;
